Move `normalize_directory_path` from airflow core to Google provider package
